refactor: extract config types
